+++ services/morningService.js
@@ -1,29 +1,31 @@
 import { executeQuery } from '../database/database.js'
-import { getToday } from '../utils.js'
 
-const getMorningReports= async() => {
+const getReports= async() => {
     const res = await executeQuery("SELECT * FROM morning_reports");
     return res.rowsOfObjects(); 
 }
 
-const getTodaysMorningReport = async() => {
-    const res = await executeQuery("SELECT * FROM morning_reports WHERE date=$1", getToday())
+const getReportByDate = async(date) => {
+    const res = await executeQuery("SELECT * FROM morning_reports WHERE date=$1", date);
     if (res) {
         return res.rowsOfObjects()
     }
     return []
 }
 
-const addMorningReport = async(report) => {
-    await executeQuery('DELETE FROM morning_reports WHERE date = $1', report.date)
+
+const addReport = async(report) => {
+    await executeQuery('DELETE FROM morning_reports WHERE date = $1 AND user_id = $2', report.date, report.userId)
     await executeQuery(
         'INSERT INTO morning_reports (sleep_duration, sleep_quality, mood, date, user_id) VALUES ($1, $2, $3, $4, $5)',
         report.sleepDuration, report.sleepQuality, report.mood, report.date, report.userId
     )
 }
 
+
+
 export { 
-    getMorningReports, 
-    addMorningReport,
-    getTodaysMorningReport
+    getReports, 
+    addReport,
+    getReportByDate
 };
